requirements.txt
google-genai
pandas
streamlit
plotly
python-dotenv
anvil-uplink
